requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
httpx==0.25.2
apscheduler==3.10.4
pydantic>=2.9.0
pydantic-settings>=2.5.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-api-python-client==2.108.0
Pillow>=10.2.0
pyyaml==6.0.1
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-jose[cryptography]==3.3.0
aiofiles==23.2.1
